Improve settings architecture
